refactor(scripts): rename src/format to src/jsp
Because soon, in service of:

https://github.com/liferay/liferay-npm-tools/issues/309

the code in this directory will be doing more that just formatting JSP;
it will also be linting it.
