+++ packages/liferay-npm-scripts/__fixtures__/jsp/configuration.jsp
