+++ packages/liferay-npm-scripts/__fixtures__/jsp/details.jsp
