+++ packages/liferay-npm-scripts/__fixtures__/jsp/edit_content_redirect.jsp
