+++ packages/liferay-npm-scripts/__fixtures__/jsp/page.jsp
