+++ packages/liferay-npm-scripts/__fixtures__/jsp/recaptcha.jsp
