+++ packages/liferay-npm-scripts/__fixtures__/jsp/roles.jsp
