+++ packages/liferay-npm-scripts/__fixtures__/jsp/source.jsp
