+++ packages/liferay-npm-scripts/__fixtures__/jsp/view.jsp
