+++ packages/liferay-npm-scripts/__fixtures__/jsp/view_calendar_menus.jspf
