+++ packages/liferay-npm-scripts/src/jsp/Lexer.js
