+++ packages/liferay-npm-scripts/src/jsp/extractJS.js
