+++ packages/liferay-npm-scripts/src/jsp/formatJSP.js
