+++ packages/liferay-npm-scripts/src/jsp/getPaddedReplacement.js
