+++ packages/liferay-npm-scripts/src/jsp/isJSP.js
