+++ packages/liferay-npm-scripts/src/jsp/lex.js
