+++ packages/liferay-npm-scripts/src/jsp/padLines.js
