+++ packages/liferay-npm-scripts/src/jsp/restoreTags.js
