+++ packages/liferay-npm-scripts/src/jsp/stripIndents.js
