+++ packages/liferay-npm-scripts/src/jsp/tagReplacements.js
