+++ packages/liferay-npm-scripts/src/jsp/trim.js
